requirements.txt
fastapi
uvicorn
requests
streamlit
python-dotenv
google-generativeai
redis
pymongo